requirements.txt
Flask==3.0.0
pandas==2.2.0
xlsxwriter==3.1.9
Werkzeug==3.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3 
python-dotenv==1.0.0
Flask-WTF==1.2.1
WTForms==3.1.1
schedule==1.2.0
gunicorn==21.2.0 